+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.9.1] - 2024-01-07
+
 ### Added
 
 - Add default map from `System.Object` to `any`
@@ -158,7 +160,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.9.0...HEAD
+[unreleased]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.9.1...HEAD
+[0.9.1]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.9.1...v0.9.0
 [0.9.0]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.9.0...v0.8.1
 [0.8.1]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.7.0...v0.8.0